+++ tests/lockup/programs/lockup/src/lib.rs
@@ -1,6 +1,7 @@
 //! A relatively advanced example of a lockup program. If you're new to Anchor,
 //! it's suggested to start with the other examples.
 
+use anchor_lang::accounts::state::ProgramState;
 use anchor_lang::prelude::*;
 use anchor_lang::solana_program;
 use anchor_lang::solana_program::instruction::Instruction;